requirements.txt
google-adk>=1.0.0
feedparser>=6.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
google-cloud-storage>=2.0.0
google-cloud-modelarmor>=0.1.0
